requirements.txt
# Unified Scraper Framework Dependencies

# Core scraping
playwright>=1.40.0
lz4>=4.3.2

# Database
psycopg2-binary>=2.9.9

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
